requirements.txt
requests>=2.28.0
html2text>=2020.1.16
python-dotenv>=1.0.0
